Update: move `RegionSplitter` to it's own module
